requirements.txt
python-arango
streamlit
python-dotenv